Added loggers and introduced some checks

+++ bank-api/src/main/java/com/bank/bankapi/resources/EmployeeResources.java
@@ -7,6 +7,9 @@ import com.bank.bankapi.services.EmployeeService;
 import com.bank.bankapi.util.GenerateJWTToken;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
+import lombok.extern.slf4j.Slf4j;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -23,7 +26,7 @@ import java.util.Map;
 @RestController
 @RequestMapping("/api/employee")
 public class EmployeeResources {
-
+    Logger logger= LoggerFactory.getLogger(EmployeeResources.class);
     @Autowired
     EmployeeService employeeService;
     @Autowired
@@ -41,19 +44,28 @@ public class EmployeeResources {
     public ResponseEntity<Map<String,String>> registerEmployee(@RequestBody Map<String,Object> data, HttpServletRequest request){
         int userID= (Integer) request.getAttribute("userId");
         Employee employeeAuth =  employeeRepository.findEmployeeById(userID);
+        Map<String, String> map= new HashMap<>();
         if(employeeAuth==null ||employeeAuth.getRole() != Employee.Role.ADMIN || employeeAuth.is_active() == false)
         {
-            Map<String, String> map= new HashMap<>();
+
             map.put("message","You are not authorized to do this function");
+            logger.error("Employee is not authorized/logged out with id {}",userID);
             return new ResponseEntity<>(map,HttpStatus.UNAUTHORIZED);
         }
         String firstName = (String)data.get("firstName");
         String lastName= (String)data.get("lastName");
         String phone= (String)data.get("phone");
         String password=(String)data.get("password");
+        if(firstName==null||lastName==null||phone==null||password==null||firstName.isBlank()||lastName.isBlank()||phone.isBlank()||password.isBlank())
+        {
+            logger.error("Data not present");
+            map.put("message","Data is absent");
+            return new ResponseEntity<>(map,HttpStatus.BAD_REQUEST);
+
+        }
         Employee employee= employeeService.registerEmployee(firstName,lastName,password, Employee.Role.EMPLOYEE,phone);
-        Map<String ,String > map = new HashMap<>();
-        map.put("message","Employee has been registered");
+        logger.info("Employee registered with id {}",employee.getUser_id());
+        map.put("id", String.valueOf(employee.getUser_id()));
         return new ResponseEntity<>(map, HttpStatus.OK);
 
     }
@@ -68,6 +80,7 @@ public class EmployeeResources {
 
         String phone= (String)data.get("phone");
         String password=(String)data.get("password");
+        logger.info("Login :Details received phone {} ",phone);
         Employee employee= employeeService.validateEmployee(phone,password,true);
 
         return new ResponseEntity<>(generateJWTToken.generateJWTToken(employee), HttpStatus.OK);
@@ -84,13 +97,15 @@ public class EmployeeResources {
         int userID= (Integer) request.getAttribute("userId");
         Employee employeeAuth =  employeeRepository.findEmployeeById(userID);
         Map<String, String> map = new HashMap<>();
-        if(employeeAuth==null ||employeeAuth.getRole() != Employee.Role.ADMIN || employeeAuth.is_active() == false)
+        if(employeeAuth==null || employeeAuth.is_active() == false)
         {
             map.put("message","You are not authorized to do this function");
+            logger.error("Employee is not authorized/logged out with id {}",userID);
             return new ResponseEntity<>(map,HttpStatus.UNAUTHORIZED);
         }
 
         Employee employee= employeeService.logout(employeeAuth);
+        logger.info("Logout Successfull for employee {}",employee.getUser_id());
         map.put("message","Employee has been logged out");
         return new ResponseEntity<>(map, HttpStatus.OK);
 
@@ -110,15 +125,24 @@ public class EmployeeResources {
         if(employeeAuth==null ||employeeAuth.getRole() != Employee.Role.ADMIN || employeeAuth.is_active() == false)
         {
             map.put("message","You are not authorized to do this function");
+            logger.error("Employee is not authorized/logged out with id {}",userID);
             return new ResponseEntity<>(map,HttpStatus.UNAUTHORIZED);
         }
         String phone= (String)data.get("phone");
+        if(phone==null||phone.isBlank())
+        {
+            logger.error("Data not present");
+            map.put("message","Data is absent");
+            return new ResponseEntity<>(map,HttpStatus.BAD_REQUEST);
+        }
         boolean flag= employeeService.deleteEmployee(phone);
         if(flag) {
+            logger.info("Employee with phone {} deleted successfully",phone);
             map.put("message", "Employee Deleted Successfully");
             return new ResponseEntity<>(map, HttpStatus.OK);
         }
         else {
+            logger.error("Unable to delete the Employee with phone {}",phone);
             map.put("message", "Unable to Delete the Employee");
             return new ResponseEntity<>(map, HttpStatus.BAD_REQUEST);
         }
